requirements.txt
argparse
pytube
opencv-python
numpy
requests
num2words
hurry
progress
ffmpeg-python